model-services/requirements.txt
Click==7.0
colorclass==2.2.0
homer-text==0.4.1
nltk==3.4.1
pika==1.2.0
Pyphen==0.9.5
repoze.lru==0.7
six==1.12.0
terminaltables==3.1.0
textstat==0.5.6
